% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr1package.R
\docType{package}
\name{dpr1_package}
\alias{dpr1package}
\alias{dpr1package-package}
\alias{dpr1_package}
\alias{dpr1_package-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
dpr1package
A data package for dpr1package.

Use \code{data(package='dpr1_package')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
