% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr1package.R
\docType{data}
\name{iris}
\alias{iris}
\title{Descriptive data title}
\format{
a \code{data.frame} containing the following fields:
\describe{
\item{Sepal.Length}{}
\item{Sepal.Width}{}
\item{Petal.Length}{}
\item{Petal.Width}{}
\item{Species}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
