% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{Descriptive data title}
\format{
a \code{data.table} containing the following fields:
\describe{
\item{mpg}{}
\item{cyl}{}
\item{disp}{}
\item{hp}{}
\item{drat}{}
\item{wt}{}
\item{qsec}{}
\item{vs}{}
\item{am}{}
\item{gear}{}
\item{carb}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{dpr.git.collision}
}
