% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpr1package.R
\docType{data}
\name{mtcars_mod}
\alias{mtcars_mod}
\title{Descriptive data title}
\format{
a \code{data.frame} containing the following fields:
\describe{
\item{mpg}{}
\item{cyl}{}
\item{disp}{}
\item{hp}{}
\item{drat}{}
\item{wt}{}
\item{qsec}{}
\item{vs}{}
\item{am}{}
\item{gear}{}
\item{carb}{}
\item{test}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
