% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R, R/dpr1_package.R,
%   R/dpr1package.R
\docType{data}
\name{trees}
\alias{trees}
\title{Descriptive data title}
\format{
a \code{data.table} containing the following fields:
\describe{
\item{Girth}{}
\item{Height}{}
\item{Volume}{}
}

a \code{data.frame} containing the following fields:
\describe{
\item{y}{}
\item{Girth}{}
\item{Height}{}
\item{Volume}{}
\item{x}{}
}

a \code{data.frame} containing the following fields:
\describe{
\item{y}{}
\item{Girth}{}
\item{Height}{}
\item{Volume}{}
\item{x}{}
}
}
\source{
The data comes from________________________.

The data comes from________________________.

The data comes from________________________.
}
\description{
Detailed description of the data

Detailed description of the data

Detailed description of the data
}
\seealso{
\link{dpr.git.collision}
}
